// src/types/EnvType.ts
export type EnvType = {
  NEXT_PUBLIC_APPWRITE_URL: string;
  NEXT_PUBLIC_APPWRITE_PROJECT_ID: string;
  NEXT_PUBLIC_APPWRITE_DATABASE_ID: string;
  NEXT_PUBLIC_APPWRITE_COLLECTION_ID: string;
  NEXT_PUBLIC_APPWRITE_BUCKET_ID: string;
  NEXT_PUBLIC_TINY_MCE_API_KEY: string;
  NEXT_PUBLIC_APPWRITE_API_KEY: string;
};